Revert "New version (#1)"
This reverts commit ee917f12b617f156bd368ce9966a34a54323729c.
